remove shorthand from world config input
