+++ packages/world/ts/config/v2/world.test.ts
@@ -67,50 +67,6 @@ describe("resolveWorldConfig", () => {
           type: "table",
         },
       },
-      namespaces: {
-        ExampleNamespace: {
-          tables: {
-            ExampleTable: {
-              tableId: resourceToHex({ type: "table", namespace: "ExampleNamespace", name: "ExampleTable" }),
-              schema: {
-                id: {
-                  type: "address",
-                  internalType: "address",
-                },
-                value: {
-                  type: "uint256",
-                  internalType: "uint256",
-                },
-                dynamic: {
-                  type: "string",
-                  internalType: "string",
-                },
-              },
-              keySchema: {
-                id: {
-                  type: "address",
-                  internalType: "address",
-                },
-              },
-              valueSchema: {
-                value: {
-                  type: "uint256",
-                  internalType: "uint256",
-                },
-                dynamic: {
-                  type: "string",
-                  internalType: "string",
-                },
-              },
-              key: ["id"],
-              name: "ExampleTable",
-              namespace: "ExampleNamespace",
-              codegen: { ...TABLE_CODEGEN_DEFAULTS, dataStruct: true as boolean },
-              type: "table",
-            },
-          },
-        },
-      },
       userTypes: {},
       enums: {},
       codegen: CODEGEN_DEFAULTS,
@@ -188,50 +144,6 @@ describe("resolveWorldConfig", () => {
           type: "table",
         },
       },
-      namespaces: {
-        ExampleNamespace: {
-          tables: {
-            ExampleTable: {
-              tableId: resourceToHex({ type: "table", namespace: "ExampleNamespace", name: "ExampleTable" }),
-              schema: {
-                id: {
-                  type: "address",
-                  internalType: "Static",
-                },
-                value: {
-                  type: "uint8",
-                  internalType: "MyEnum",
-                },
-                dynamic: {
-                  type: "string",
-                  internalType: "Dynamic",
-                },
-              },
-              keySchema: {
-                id: {
-                  type: "address",
-                  internalType: "Static",
-                },
-              },
-              valueSchema: {
-                value: {
-                  type: "uint8",
-                  internalType: "MyEnum",
-                },
-                dynamic: {
-                  type: "string",
-                  internalType: "Dynamic",
-                },
-              },
-              key: ["id"],
-              name: "ExampleTable",
-              namespace: "ExampleNamespace",
-              codegen: { ...TABLE_CODEGEN_DEFAULTS, dataStruct: true as boolean },
-              type: "table",
-            },
-          },
-        },
-      },
       userTypes: {
         Static: { type: "address", filePath: "path/to/file" },
         Dynamic: { type: "string", filePath: "path/to/file" },
@@ -299,254 +211,6 @@ describe("resolveWorldConfig", () => {
   });
 
   describe("should have the same output as `resolveWorldConfig` for store config inputs", () => {
-    it("should accept a shorthand store config as input and expand it", () => {
-      const config = resolveWorldConfig({ tables: { Name: "address" } });
-      const expected = {
-        tables: {
-          Name: {
-            tableId: resourceToHex({ type: "table", namespace: "", name: "Name" }),
-            schema: {
-              id: {
-                type: "bytes32",
-                internalType: "bytes32",
-              },
-              value: {
-                type: "address",
-                internalType: "address",
-              },
-            },
-            keySchema: {
-              id: {
-                type: "bytes32",
-                internalType: "bytes32",
-              },
-            },
-            valueSchema: {
-              value: {
-                type: "address",
-                internalType: "address",
-              },
-            },
-            key: ["id"],
-            name: "Name",
-            namespace: "",
-            codegen: { ...TABLE_CODEGEN_DEFAULTS, dataStruct: false as boolean },
-            type: "table",
-          },
-        },
-        userTypes: {},
-        enums: {},
-        namespace: "",
-        codegen: CODEGEN_DEFAULTS,
-        namespaces: {},
-        deployment: DEPLOYMENT_DEFAULTS,
-        ...CONFIG_DEFAULTS,
-      } as const;
-
-      attest<typeof expected>(config).equals(expected);
-      attest<typeof config>(expected);
-    });
-
-    it("should accept a user type as input and expand it", () => {
-      const config = resolveWorldConfig({
-        tables: { Name: "CustomType" },
-        userTypes: { CustomType: { type: "address", filePath: "path/to/file" } },
-      });
-      const expected = {
-        tables: {
-          Name: {
-            tableId: resourceToHex({ type: "table", namespace: "", name: "Name" }),
-            schema: {
-              id: {
-                type: "bytes32",
-                internalType: "bytes32",
-              },
-              value: {
-                type: "address",
-                internalType: "CustomType",
-              },
-            },
-            keySchema: {
-              id: {
-                type: "bytes32",
-                internalType: "bytes32",
-              },
-            },
-            valueSchema: {
-              value: {
-                type: "address",
-                internalType: "CustomType",
-              },
-            },
-            key: ["id"],
-            name: "Name",
-            namespace: "",
-            codegen: { ...TABLE_CODEGEN_DEFAULTS, dataStruct: false as boolean },
-            type: "table",
-          },
-        },
-        userTypes: { CustomType: { type: "address", filePath: "path/to/file" } },
-        enums: {},
-        namespace: "",
-        codegen: CODEGEN_DEFAULTS,
-        namespaces: {},
-        deployment: DEPLOYMENT_DEFAULTS,
-        ...CONFIG_DEFAULTS,
-      } as const;
-
-      attest<typeof expected>(config).equals(expected);
-    });
-
-    it("given a schema with a key field with static ABI type, it should use `id` as single key", () => {
-      const config = resolveWorldConfig({ tables: { Example: { id: "address", name: "string", age: "uint256" } } });
-      const expected = {
-        tables: {
-          Example: {
-            tableId: resourceToHex({ type: "table", namespace: "", name: "Example" }),
-            schema: {
-              id: {
-                type: "address",
-                internalType: "address",
-              },
-              name: {
-                type: "string",
-                internalType: "string",
-              },
-              age: {
-                type: "uint256",
-                internalType: "uint256",
-              },
-            },
-            keySchema: {
-              id: {
-                type: "address",
-                internalType: "address",
-              },
-            },
-            valueSchema: {
-              name: {
-                type: "string",
-                internalType: "string",
-              },
-              age: {
-                type: "uint256",
-                internalType: "uint256",
-              },
-            },
-            key: ["id"],
-            name: "Example",
-            namespace: "",
-            codegen: { ...TABLE_CODEGEN_DEFAULTS, dataStruct: true as boolean },
-            type: "table",
-          },
-        },
-        userTypes: {},
-        enums: {},
-        namespace: "",
-        codegen: CODEGEN_DEFAULTS,
-        namespaces: {},
-        deployment: DEPLOYMENT_DEFAULTS,
-        ...CONFIG_DEFAULTS,
-      } as const;
-
-      attest<typeof expected>(config).equals(expected);
-    });
-
-    it("given a schema with a key field with static custom type, it should use `id` as single key", () => {
-      const config = resolveWorldConfig({ tables: { Example: { id: "address", name: "string", age: "uint256" } } });
-      const expected = {
-        tables: {
-          Example: {
-            tableId: resourceToHex({ type: "table", namespace: "", name: "Example" }),
-            schema: {
-              id: {
-                type: "address",
-                internalType: "address",
-              },
-              name: {
-                type: "string",
-                internalType: "string",
-              },
-              age: {
-                type: "uint256",
-                internalType: "uint256",
-              },
-            },
-            keySchema: {
-              id: {
-                type: "address",
-                internalType: "address",
-              },
-            },
-            valueSchema: {
-              name: {
-                type: "string",
-                internalType: "string",
-              },
-              age: {
-                type: "uint256",
-                internalType: "uint256",
-              },
-            },
-            key: ["id"],
-            name: "Example",
-            namespace: "",
-            codegen: { ...TABLE_CODEGEN_DEFAULTS, dataStruct: true as boolean },
-            type: "table",
-          },
-        },
-        userTypes: {},
-        enums: {},
-        namespace: "",
-        codegen: CODEGEN_DEFAULTS,
-        namespaces: {},
-        deployment: DEPLOYMENT_DEFAULTS,
-        ...CONFIG_DEFAULTS,
-      } as const;
-
-      attest<typeof expected>(config).equals(expected);
-    });
-
-    it("throw an error if the shorthand doesn't include a key field", () => {
-      attest(() =>
-        resolveWorldConfig({
-          tables: {
-            // @ts-expect-error Invalid schema. Expected an `id` field with a static ABI type or an explicit `key` option.
-            Example: {
-              name: "string",
-              age: "uint256",
-            },
-          },
-        }),
-      ).throwsAndHasTypeError(
-        "Invalid schema. Expected an `id` field with a static ABI type or an explicit `key` option.",
-      );
-    });
-
-    it("throw an error if the shorthand config includes a non-static key field", () => {
-      attest(() =>
-        // @ts-expect-error Invalid schema. Expected an `id` field with a static ABI type or an explicit `key` option.
-        resolveWorldConfig({ tables: { Example: { id: "string", name: "string", age: "uint256" } } }),
-      ).throwsAndHasTypeError(
-        "Invalid schema. Expected an `id` field with a static ABI type or an explicit `key` option.",
-      );
-    });
-
-    it("throw an error if the shorthand config includes a non-static user type as key field", () => {
-      attest(() =>
-        resolveWorldConfig({
-          // @ts-expect-error Invalid schema. Expected an `id` field with a static ABI type or an explicit `key` option.
-          tables: { Example: { id: "dynamic", name: "string", age: "uint256" } },
-          userTypes: {
-            dynamic: { type: "string", filePath: "path/to/file" },
-            static: { type: "address", filePath: "path/to/file" },
-          },
-        }),
-      ).throwsAndHasTypeError(
-        "Invalid schema. Expected an `id` field with a static ABI type or an explicit `key` option.",
-      );
-    });
-
     it("should return the full config given a full config with one key", () => {
       const config = resolveWorldConfig({
         tables: {
@@ -601,7 +265,6 @@ describe("resolveWorldConfig", () => {
         enums: {},
         namespace: "",
         codegen: CODEGEN_DEFAULTS,
-        namespaces: {},
         deployment: DEPLOYMENT_DEFAULTS,
         ...CONFIG_DEFAULTS,
       } as const;
@@ -670,7 +333,6 @@ describe("resolveWorldConfig", () => {
         enums: {},
         namespace: "",
         codegen: CODEGEN_DEFAULTS,
-        namespaces: {},
         deployment: DEPLOYMENT_DEFAULTS,
         ...CONFIG_DEFAULTS,
       } as const;
@@ -731,7 +393,6 @@ describe("resolveWorldConfig", () => {
           enums: {},
           namespace: "",
           codegen: CODEGEN_DEFAULTS,
-          namespaces: {},
           deployment: DEPLOYMENT_DEFAULTS,
           ...CONFIG_DEFAULTS,
         } as const;
@@ -835,7 +496,6 @@ describe("resolveWorldConfig", () => {
         enums: {},
         namespace: "",
         codegen: CODEGEN_DEFAULTS,
-        namespaces: {},
         deployment: DEPLOYMENT_DEFAULTS,
         ...CONFIG_DEFAULTS,
       } as const;
@@ -860,6 +520,7 @@ describe("resolveWorldConfig", () => {
           Dynamic: { type: "string", filePath: "path/to/file" },
         },
       });
+
       const expected = {
         tables: {
           First: {
@@ -946,7 +607,6 @@ describe("resolveWorldConfig", () => {
         enums: {},
         namespace: "",
         codegen: CODEGEN_DEFAULTS,
-        namespaces: {},
         deployment: DEPLOYMENT_DEFAULTS,
         ...CONFIG_DEFAULTS,
       } as const;
@@ -1075,7 +735,6 @@ describe("resolveWorldConfig", () => {
         },
         namespace: "",
         codegen: CODEGEN_DEFAULTS,
-        namespaces: {},
         deployment: DEPLOYMENT_DEFAULTS,
         ...CONFIG_DEFAULTS,
       } as const;
@@ -1097,9 +756,10 @@ describe("resolveWorldConfig", () => {
 
     it("should extend the output Config type", () => {
       const config = resolveWorldConfig({
-        tables: { Name: "CustomType" },
+        tables: { Name: { schema: { key: "CustomType" }, key: ["key"] } },
         userTypes: { CustomType: { type: "address", filePath: "path/to/file" } },
       });
+
       attest<true, typeof config extends Config ? true : false>();
     });
